Refs #35718, Refs #32179 -- Moved JSONObject to django.db.models.functions.json.
