+++ django/db/models/functions/__init__.py
@@ -1,4 +1,4 @@
-from .comparison import Cast, Coalesce, Collate, Greatest, JSONObject, Least, NullIf
+from .comparison import Cast, Coalesce, Collate, Greatest, Least, NullIf
 from .datetime import (
     Extract,
     ExtractDay,
@@ -25,6 +25,7 @@ from .datetime import (
     TruncWeek,
     TruncYear,
 )
+from .json import JSONObject
 from .math import (
     Abs,
     ACos,
@@ -97,7 +98,6 @@ __all__ = [
     "Coalesce",
     "Collate",
     "Greatest",
-    "JSONObject",
     "Least",
     "NullIf",
     # datetime
@@ -125,6 +125,8 @@ __all__ = [
     "TruncTime",
     "TruncWeek",
     "TruncYear",
+    # json
+    "JSONObject",
     # math
     "Abs",
     "ACos",
